+++ .changes/unreleased/Features-20230103-152219.yaml
@@ -1,0 +1,7 @@
+kind: Features
+body: Making timestamps optional
+time: 2023-01-03T15:22:19.38878-06:00
+custom:
+  Author: callum-mcdata
+  Issue: "200"
+  PR: "201"
