+++ integration_tests/models/materialized_models/fact_orders_duplicate.sql
@@ -1,0 +1,4 @@
+select 
+    *
+    ,round(order_total - (order_total/2)) as discount_total
+from {{ref('fact_orders_duplicate_source')}}
