+++ integration_tests/models/metric_definitions/base_average_metric.yml
@@ -4,7 +4,7 @@ metrics:
     model: ref('fact_orders')
     label: Total Discount ($)
     timestamp: order_date
-    time_grains: [day, week, month, all_time]
+    time_grains: [day, week, month, test]
     calculation_method: average
     expression: discount_total
     dimensions:
