+++ integration_tests/models/metric_definitions/base_sum_metric.yml
@@ -11,6 +11,17 @@ metrics:
       - had_discount
       - order_country
 
+  - name: base_sum_metric_duplicate
+    model: ref('fact_orders_duplicate')
+    label: Order Total ($)
+    timestamp: order_date
+    time_grains: [day, week, month, all_time]
+    calculation_method: sum
+    expression: order_total
+    dimensions:
+      - had_discount
+      - order_country
+
   - name: base_sum_metric__14_day_window
     model: ref('fact_orders')
     label: Order Total ($)
