+++ integration_tests/models/metric_testing_models/base_average_metric.sql
@@ -1,6 +1,6 @@
 select *
 from 
 {{ metrics.calculate(metric('base_average_metric'), 
-    grain='month', 
+    grain='test', 
     dimensions=['had_discount']) 
 }}
