+++ integration_tests/models/metric_testing_models/base_sum_metric.sql
@@ -1,6 +1,9 @@
 select *
 from 
 {{ metrics.calculate(
-    [metric('base_sum_metric'),metric('base_test_metric')], 
-    dimensions=['had_discount','is_weekend']) 
-}}
+    [metric('base_sum_metric'), metric('base_sum_metric_duplicate')],
+    secondary_calculations=[
+        metrics.period_over_period(comparison_strategy="difference", interval=1, alias = "1mth")
+    ]
+    )
+    }}
