+++ integration_tests/models/metric_testing_models/develop_metric_no_timestamp.sql
@@ -1,0 +1,22 @@
+{% set my_metric_yml -%}
+{% raw %}
+
+metrics:
+  - name: develop_metric
+    model: ref('fact_orders')
+    label: Total Discount ($)
+    calculation_method: average
+    expression: discount_total
+    dimensions:
+      - had_discount
+      - order_country
+
+{% endraw %}
+{%- endset %}
+
+select * 
+from {{ metrics.develop(
+        develop_yml=my_metric_yml,
+        metric_list=['develop_metric'],
+        dimensions=['order_country'])
+    }}
