+++ integration_tests/seeds/source/fact_orders_source.csv
@@ -6,6 +6,6 @@ order_id,order_country,order_total,had_discount,customer_id,order_date
 5,Portugal,1,false,4,01/08/2022
 6,China,1,false,5,01/21/2022
 7,Germany,1,true,2,01/22/2022
-8,Greenland,0,true,1,02/15/2022
+8,Greenland,1,true,1,02/15/2022
 9,Bangladesh,1,false,2,02/03/2022
 10,Sweden,1,false,3,02/13/2022
