+++ macros/sql_gen/gen_joined_metrics_cte.sql
@@ -23,47 +23,49 @@
 
     select
         {% if grain -%}
-        date_{{grain}}
-        {%- else -%}
-        1 as comma_placeholder
+        date_{{grain}},
         {%- endif -%}
         
         {%- for calendar_dim in calendar_dimensions %}
-        , coalesce(
+        coalesce(
             {%- for metric_name in metric_tree.parent_set %}
                 {{metric_name}}__final.{{ calendar_dim }}{%- if not loop.last -%},{% endif %}
                 {%- if metric_tree.parent_set | length == 1 -%}
                 , NULL
                 {%- endif -%}
             {% endfor %}
-            ) as {{calendar_dim}}
+            ) as {{calendar_dim}},
         {% endfor %}
 
     {%- for period in relevant_periods %}
-        , coalesce(
+        coalesce(
         {%- for metric_name in metric_tree.parent_set %}
             {{metric_name}}__final.date_{{ period }} {%- if not loop.last -%},{% endif %}
             {%- if metric_tree.parent_set | length == 1 %}
             , NULL
             {%- endif -%}
         {% endfor %}
-        ) as date_{{period}}
+        ) as date_{{period}},
     {%- endfor %}
 
 
     {%- for dim in dimensions %}
-        , coalesce(
+        coalesce(
         {%- for metric_name in metric_tree.parent_set %}
             {{metric_name}}__final.{{ dim }} {%- if not loop.last -%},{% endif %}
             {%- if metric_tree.parent_set | length == 1 %}
             , NULL
             {%- endif -%}
         {% endfor %}
-        ) as {{dim}}
+        ) as {{dim}},
     {%- endfor %}
 
     {%- for metric_name in metric_tree.parent_set %}
-        , {{metric_name}} as {{metric_name}}
+        {%- if not metrics_dictionary[metric_name].config.get("treat_null_values_as_zero", True) %}
+        {{metric_name}} as {{metric_name}} {%- if not loop.last -%}, {%- endif -%}
+        {%- else  %}  
+        coalesce({{metric_name}},0) as {{metric_name}} {%- if not loop.last -%}, {%- endif -%}
+        {%- endif %}  
     {%- endfor %}  
 
     from 
@@ -152,37 +154,24 @@
     select 
 
     {%- if grain %}
-        date_{{grain}}
+        date_{{grain}},
     {% endif -%}
 
     {%- for period in relevant_periods %}
-        ,date_{{ period }}
+        date_{{ period }},
     {%- endfor %}
 
     {%- for calendar_dim in calendar_dimensions %}
-        , {{ calendar_dim }}
+        {{ calendar_dim }},
     {%- endfor %}
 
     {%- for dim in dimensions %}
-        , {{ dim }}
+        {{ dim }},
     {%- endfor %}
 
-        {%- for metric_name in metric_tree.parent_set %}
-        , {{metric_name}}
-        {%- endfor %}  
-        {%- for metric in metric_tree.expression_set %}
-        , {{ metric }}
-        {% endfor -%}
-
-        {%- for metric_name in metric_tree.parent_set %}
-        {%- if loop.first -%}, {%- endif -%}
-        {{metric_name}}
-        {%- endfor %}  
-        {%- for metric in metric_tree.expression_set %}
-        , {{ metric }}
-
-
-    {%- endif -%}
+    {%- for metric in metric_tree.parent_set|list + metric_tree.expression_set|list %}
+        {{metric}}{%- if not loop.last -%}, {%- endif -%}
+    {%- endfor %}  
 
     {% if metric_tree.expression_set | length == 0 %}
     from first_join_metrics
