+++ macros/validation/validate_grain.sql
@@ -1,4 +1,4 @@
-{% macro validate_grain(grain, metric_tree, metrics_dictionary, secondary_calculations) %}
+{% macro validate_grain(grain, metric_tree, metrics_dictionary, secondary_calculations, dimensions) %}
 
     {# We loop through the full set here to ensure that the provided grain works for all metrics
     returned or used, not just those listed #}
@@ -8,7 +8,7 @@
         {%- endif -%}
 
         {% for metric_name in metric_tree.full_set %}
-            {% set metric_relation = metric(metric_name)%}
+            {% set metric_relation = metrics_dictionary[metric_name]%}
             {% if grain not in metric_relation.time_grains%}
                 {% if metric_name not in metric_tree.base_set %}
                     {%- do exceptions.raise_compiler_error("The metric " ~ metric_name ~ " is an upstream metric of one of the provided metrics. The grain " ~ grain ~ " is not defined in its metric definition.") %}
@@ -18,19 +18,18 @@
             {% endif %}
         {% endfor %}
 
-        {% if not grain %}
+    {% elif not grain %}
 
-            {% if secondary_calculations | length > 0 %}
-                {%- do exceptions.raise_compiler_error("Aggregating without a grain does not support secondary calculations.") %}
-            {% endif %}
-
-            {% for metric_name in metric_tree.full_set %}
-                {% if metrics_dictionary[metric_name].window is not none%}
-                    {%- do exceptions.raise_compiler_error("Aggregating without a grain does not support metrics with window definitions.") %}
-                {% endif%}
-            {% endfor%}
-
+        {% if secondary_calculations | length > 0 %}
+            {%- do exceptions.raise_compiler_error("Using secondary calculations without a grain is not supported.") %}
         {% endif %}
+
+        {% for metric_name in metric_tree.full_set %}
+            {% if metrics_dictionary[metric_name].window is not none%}
+                {%- do exceptions.raise_compiler_error("Aggregating without a grain does not support metrics with window definitions.") %}
+            {% endif%}
+        {% endfor%}
+
     {% endif %}
 
 {% endmacro %}
