--- tests/functional/metric_options/date_grains/test_month_grain.py
@@ -1,116 +1,0 @@
-from struct import pack
-import os
-import pytest
-from dbt.tests.util import run_dbt
-
-# our file contents
-from tests.functional.fixtures import (
-    fact_orders_source_csv,
-    fact_orders_sql,
-    fact_orders_yml,
-)
-
-# models/month_grain_metric.sql
-month_grain_metric_sql = """
-select *
-from 
-{{ metrics.calculate(metric('month_grain_metric'), 
-    grain='month'
-    )
-}}
-"""
-
-# models/month_grain_metric.yml
-month_grain_metric_yml = """
-version: 2 
-models:
-  - name: month_grain_metric
-    tests: 
-      - metrics.metric_equality:
-          compare_model: ref('month_grain__expected')
-metrics:
-  - name: month_grain_metric
-    model: ref('fact_orders')
-    label: Total Discount ($)
-    timestamp: order_date
-    time_grains: [day, week, month]
-    calculation_method: count
-    expression: order_total
-    dimensions:
-      - had_discount
-      - order_country
-"""
-
-# seeds/month_grain__expected.csv
-month_grain__expected_csv = """
-date_month,month_grain_metric
-2022-02-01,3
-2022-01-01,7
-""".lstrip()
-
-class TestmonthGrain:
-
-    # configuration in dbt_project.yml
-    # setting bigquery as table to get around query complexity 
-    # resource constraints with compunding views
-    if os.getenv('dbt_target') == 'bigquery':
-        @pytest.fixture(scope="class")
-        def project_config_update(self):
-            return {
-            "name": "example",
-            "models": {"+materialized": "table"}
-            }
-    else: 
-        @pytest.fixture(scope="class")
-        def project_config_update(self):
-            return {
-            "name": "example",
-            "models": {"+materialized": "view"}
-            }  
-    # install current repo as package
-    @pytest.fixture(scope="class")
-    def packages(self):
-        return {
-            "packages": [
-                {"local": os.getcwd()}
-                ]
-        }
-
-
-    # everything that goes in the "seeds" directory
-    @pytest.fixture(scope="class")
-    def seeds(self):
-        return {
-            "fact_orders_source.csv": fact_orders_source_csv,
-            "month_grain__expected.csv": month_grain__expected_csv,
-        }
-
-    # everything that goes in the "models" directory
-    @pytest.fixture(scope="class")
-    def models(self):
-        return {
-            "fact_orders.sql": fact_orders_sql,
-            "fact_orders.yml": fact_orders_yml,
-            "month_grain_metric.sql": month_grain_metric_sql,
-            "month_grain_metric.yml": month_grain_metric_yml
-        }
-
-    def test_build_completion(self,project,):
-        # running deps to install package
-        results = run_dbt(["deps"])
-
-        # seed seeds
-        results = run_dbt(["seed"])
-        assert len(results) == 2
-
-        # initial run
-        results = run_dbt(["run"])
-        assert len(results) == 3
-
-        # test tests
-        results = run_dbt(["test"]) # expect passing test
-        assert len(results) == 1
-
-        # # # validate that the results include pass
-        result_statuses = sorted(r.status for r in results)
-        assert result_statuses == ["pass"]
